requirements.txt
torch
transformers
safetensors
huggingface_hub
python-dotenv